Rag_project_stremlit_quad/requirements.txt
qdrant-client==1.9.1
openai==1.30.1
tiktoken==0.7.0
streamlit==1.35.0
python-dotenv==1.0.1
pypdf